+++ CHANGES.rst
@@ -189,3 +189,4 @@ v1.1.7
 
 v1.1.8[unreleased]
 -----------
+* Add unspecified gender option