+++ CHANGES.rst
@@ -193,3 +193,4 @@ v1.1.8[unreleased]
 * Stop requiring email on password recovery
 * On User#profile, test if cached value is instance of the profile class
 * Add public in CurrentUserSerializer
+* Remove User.PUT, use User.PATCH